requirements.txt
Pillow
qrcode
Flask
Flask-Cors